Fix shader programs
Issue was with ownership of the shaders, which would get destroyed when trying to transfer ownership to the shader program. This is because I misunderstood C++'s moves, which are not destructive moves like Rust's, and will always invoke their destructor..

Hence why I replaced them with `unique_ptr`s, and it now works.