Restructure
Now shaders are stored in a new 'release' folder which replaces the build folder. This is so you don't need to access the src folder accessing files outside of the build which is not a safe precedent to set nor is it clean.

clangd didn't seem to like that so I have added a custom vscode setting to change the directory, maybe there's a generic solution somewhere else I'll do later.