Asset2D progress
Going to make Asset2D store a simple index into a texture::Manager rather than the data and/or a path to the filesystem.

This will mean that everything is resolved upon creation of the asset, and hence there's no reason to search the filesystem for assets.

This should work fine for obj files, which could store texture indices as a map from paths, and then resolve those still at asset creation time.

This will mean that the loaders are going to need to take the texture manager, either as another parameter or as a field in a class, turning them into instances. I think a parameter would be better considering we have two separate texture managers for 2d and 3d.