Change the way transforms are used
Previously, both GlobalTransform and the 3 separate components (Position/Scale/Rotation) were supposed to be able to be mutated.

That doesn't really work out well - so now only the 3 separate components are intended to be mutated, and the GlobalTransform should be treated as immutable. So now you can rely on the separate components being updated as they are sources of truth on their own (being automatically combined into GlobalTransform, by the transform system / plugin)