Try to copy over dlls
I think CRT statically linked as I didn't get that error this time. Still getting errors with using fastgltf and glfw3 now, though. If I just need to provide the dlls for those I'll be fine with that for a release and leave it to the future to statically link everything.